# docs/openapi/admin.json
{
  "openapi": "3.2.0",
...
  "info": {
    "title": "LiteLLM-RS Admin Control Plane API",
    "version": "0.6.0",
    "summary": "Authenticated control-plane HTTP contract for the self-hosted LiteLLM-RS gateway",
    "description": "Auth, keys, teams, budgets, routing inventory, request ledger, provider configuration, routing policy, and cache administration. This document is served at GET /admin/openapi.json and requires an Admin-role caller when authentication is enabled. The stable inference contract remains GET /openapi.json and is intentionally separate. MCP, A2A, Realtime, and dashboard static assets are excluded. Response schemas omit api_key values, Authorization headers, prompt/body payloads, header maps, and raw secrets.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "The current gateway origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/admin/openapi.json": {
      "get": {
        "operationId": "getAdminOpenApi",
        "summary": "Get the admin control-plane OpenAPI document",
        "description": "Requires an authenticated Admin-role user when JWT or API-key auth is enabled. API keys without an admin user context receive 403.",
        "tags": ["OpenAPI"],
        "responses": {
          "200": {
            "description": "OpenAPI 3.2 document",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/cache": {
      "get": {
        "operationId": "getAdminCache",
        "summary": "Inspect response-cache status",
        "description": "Requires Admin role. Returns 501 when the cache is not wired into request handling.",
        "tags": ["Cache"],
        "responses": {
          "200": {
            "description": "Cache is wired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "501": {
            "description": "Cache is not wired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/cache/status": {
      "get": {
        "operationId": "getAdminCacheStatus",
        "summary": "Inspect response-cache status (alias)",
        "description": "Same payload as GET /admin/cache. Requires Admin role.",
        "tags": ["Cache"],
        "responses": {
          "200": {
            "description": "Cache is wired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "501": {
            "description": "Cache is not wired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/cache/clear": {
      "post": {
        "operationId": "clearAdminCache",
        "summary": "Clear the response cache",
        "description": "Requires Admin role. Returns 501 when the cache is not wired.",
        "tags": ["Cache"],
        "responses": {
          "200": {
            "description": "Cache cleared",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "501": {
            "description": "Cache is not wired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CacheAdminResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/providers": {
      "get": {
        "operationId": "listAdminProviders",
        "summary": "List configured providers",
        "description": "Requires Admin role. Returns public provider fields only. Raw api_key values and header maps are never serialized; api_key_ref is an env-var name when known.",
        "tags": ["Providers"],
        "responses": {
          "200": {
            "description": "Provider list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      },
      "post": {
        "operationId": "createAdminProvider",
        "summary": "Create a provider",
        "description": "Requires Admin role. api_key must be an existing env reference of the form ${VAR}. Raw secrets are rejected with 400.",
        "tags": ["Providers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProviderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderMutationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/AdminError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/AdminError" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/providers/{name}": {
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "minLength": 1 }
        }
      ],
      "patch": {
        "operationId": "updateAdminProvider",
        "summary": "Update a provider",
        "description": "Requires Admin role. Optional api_key must be ${VAR}. Responses never include api_key or headers.",
        "tags": ["Providers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProviderPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderMutationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/AdminError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/AdminError" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      },
      "delete": {
        "operationId": "deleteAdminProvider",
        "summary": "Delete a provider",
        "description": "Requires Admin role. Conflicts when the provider is still referenced by live routing or unique alias targets.",
        "tags": ["Providers"],
        "responses": {
          "200": {
            "description": "Provider deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderMutationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/AdminError" },
          "409": { "$ref": "#/components/responses/AdminError" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/request-ledger": {
      "get": {
        "operationId": "listAdminRequestLedger",
        "summary": "Query the metadata-only request ledger",
        "description": "Requires Admin role. Cursor pagination over finished requests. Prompt, body, and credentials are not fields.",
        "tags": ["Ledger"],
        "parameters": [
          {
            "name": "finished_after",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 lower bound on finished_at"
          },
          {
            "name": "finished_before",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 exclusive upper bound on finished_at"
          },
          {
            "name": "request_id",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "model",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "terminal_status",
            "in": "query",
            "schema": { "type": "string", "enum": ["completed", "failed", "cancelled"] }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque cursor from next_cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ledger page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RequestLedgerPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/AdminError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
    "/admin/request-ledger/{request_id}/reconciliation": {
      "post": {
        "operationId": "reconcileAdminRequestLedger",
        "summary": "Record supplier-verified cost and evidence for a ledger request",
        "description": "Requires Admin role. Saves evidence without changing budget counters. Identical submissions are idempotent; conflicting verification returns 409.",
        "tags": ["Ledger"],
        "parameters": [{ "name": "request_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["verified_actual_cost", "evidence_reference"],
            "properties": {
              "verified_actual_cost": { "type": "number", "minimum": 0 },
              "evidence_reference": { "type": "string", "minLength": 1, "maxLength": 512, "description": "Invoice or receipt reference; never credentials or request/response contents." }
            }
          } } }
        },
        "responses": {
          "200": { "description": "Supplier verification recorded; budget counters were not adjusted", "content": { "application/json": { "schema": {
            "type": "object", "required": ["success", "reconciliation"], "properties": {
              "success": { "type": "boolean", "const": true },
              "reconciliation": { "$ref": "#/components/schemas/RequestReconciliation" }
            }
          } } } },
          "400": { "$ref": "#/components/responses/AdminError" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Ledger request ID not found" },
          "409": { "description": "Conflicting supplier verification; existing evidence preserved" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Log out",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmptySuccess" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "default": { "$ref": "#/components/responses/AdminError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or gateway API key",
        "description": "Authorization: Bearer <jwt>, ApiKey <gw-...>, or a raw gw- prefixed key. Admin-tagged /admin/* JSON APIs additionally require an authenticated user with the Admin role (SuperAdmin satisfies has_role). API keys without that user context receive HTTP 403 with the AdminError envelope."
      }
    },
...
  }
}